requirements.txt
discord.py>=1.3.4
dateparser>=0.7.6
gitpython>=3.1.3
sqlalchemy>=1.3.18
psycopg2-binary>=2.8.5
requests>=2.24.0
graphviz>=0.14
xdot>=1.1
scipy>=1.5.1
matplotlib>=3.2.2